requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
hupper==1.12
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
PasteDeploy==3.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
psycopg2-binary==2.9.7
pydot==1.4.2
pyparsing==3.1.1
pyramid==2.0.1
pyramid-mailer==0.15.1
pyramid-rest-framework==0.0.6
pytz==2023.3
referencing==0.30.2
repoze.sendmail==4.4.1
requests==2.31.0
rest-framework-custom-viewsets==0.0.4
rest-framework-dyn-serializer==1.3.1
rest-framework-ember==1.3.2
rfc3987==1.3.8
rpds-py==0.9.2
SQLAHelper==1.0
SQLAlchemy==2.0.20
sqlalchemy-schemadisplay==1.3
sqlparse==0.4.4
strict-rfc3339==0.7
transaction==3.1.0
translationstring==1.4
typing_extensions==4.7.1
urllib3==2.0.4
venusian==3.0.0
WebOb==1.8.7
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==3.0
